# docs/openapi/unified-api.yaml
openapi: 3.1.0
info:
  title: Token Hub Unified API
  version: 0.1.0
  description: >
    Contract skeleton for developer API and console surfaces. Implementations land per phase;
    paths are stable identifiers for parallel frontend work.
servers:
  - url: https://api.example.com
    description: Production placeholder
  - url: http://localhost:8080
    description: Local gateway placeholder
  - url: http://localhost:8101
    description: user-center-service direct (dev)

tags:
  - name: Auth
  - name: Users
  - name: Models
  - name: Billing
  - name: Payments
  - name: Support
  - name: Ops

paths:
  /user/register:
    post:
      tags: [Auth, Users]
      summary: Register a new user (P1)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterRequest'
      responses:
        '200':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUserProfile'
  /user/login:
    post:
      tags: [Auth]
      summary: Password login (P1)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
      responses:
        '200':
          description: Session token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAuthToken'
  /user/logout:
    post:
      tags: [Auth]
      summary: Stateless logout hint (P1 — client discards JWT)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseEmpty'
  /user/password/forgot:
    post:
      tags: [Auth]
      summary: Request password reset code (P1 — code in server log until mail SPI wired)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Always OK when email format valid (no user enumeration)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseEmpty'
  /user/password/reset:
    post:
      tags: [Auth]
      summary: Reset password with emailed/logged code (P1)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseEmpty'
  /user/me:
    get:
      tags: [Users]
      summary: Current profile (P1)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUserProfile'
        '401':
          description: Missing or invalid JWT
  /v1/chat/completions:
    post:
      tags: [Models]
      summary: OpenAI-compatible chat (P2 non-stream, P8 stream)
      security: [{ bearerAuth: [] }, { apiKeyAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatCompletionRequest'
      responses:
        '200':
          description: Completion payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatCompletionResponse'
  /v1/models:
    get:
      tags: [Models]
      summary: List routable models (P2)
      security: [{ bearerAuth: [] }, { apiKeyAuth: [] }]
      responses:
        '200':
          description: Model catalog

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Use `Bearer sk_...` for API-key style calls routed via gateway filters.

  schemas:
    ApiResponseUserProfile:
      type: object
      description: Wrapped success envelope (aligned with ApiResponse<UserProfile>)
      properties:
        traceId:
          type: string
        code:
          type: string
        message:
          type: string
        data:
          $ref: '#/components/schemas/UserProfile'
        timestamp:
          type: string
          format: date-time

    ApiResponseAuthToken:
      type: object
      properties:
        traceId:
          type: string
        code:
          type: string
        message:
          type: string
        data:
          $ref: '#/components/schemas/AuthToken'
        timestamp:
          type: string
          format: date-time

    UserProfile:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        displayName:
          type: string

    AuthToken:
      type: object
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        expiresInSeconds:
          type: integer
          format: int64

    ApiResponseEmpty:
      type: object
      properties:
        traceId:
          type: string
        code:
          type: string
        message:
          type: string
        data:
          nullable: true
        timestamp:
          type: string
          format: date-time

    UserLoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password

    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email

    ResetPasswordRequest:
      type: object
      required: [email, code, newPassword]
      properties:
        email:
          type: string
          format: email
        code:
          type: string
        newPassword:
          type: string
          format: password

    UserRegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        displayName:
          type: string

    ChatCompletionRequest:
      type: object
      description: Mirrors OpenAI chat completion request (trimmed placeholder).
      properties:
        model:
          type: string
        messages:
          type: array
          items:
            type: object
            properties:
              role:
                type: string
              content:
                type: string

    ChatCompletionResponse:
      type: object
      description: OpenAI-compatible assistant message payload (trimmed placeholder).
      properties:
        id:
          type: string
        model:
          type: string